Move the pdc pseudo decompiler into a core plugin ##core
