+++ libr/core/p/pseudo/pseudo.h
@@ -1,0 +1,10 @@
+/* radare - LGPL - Copyright 2015-2026 - pancake */
+
+#ifndef R2_CORE_PSEUDO_H
+#define R2_CORE_PSEUDO_H
+
+#include <r_core.h>
+
+int pdc_decompile(RCore *core, const char *input);
+
+#endif
